api/requirements.txt
fastapi
fastapi[all]
gunicorn
upstash_redis
fastapi-utilities
python-dotenv
supabase
pytest
pytest-asyncio
resend
ipfs-toolkit
pydantic
s3fs
rq
